+++ voip_sip_webrtc/__manifest__.py
@@ -21,6 +21,7 @@
         'views/voip_message_compose_views.xml',
         'views/ir_actions_server_views.xml',
         'views/voip_message_template_views.xml',
+        'views/menus.xml',
         'security/ir.model.access.csv',
         'data/voip_ringtone.xml',
         'data/voip_settings.xml',