+++ voip_sip_webrtc_twilio/models/voip_call.py
@@ -8,3 +8,8 @@ class VoipCallInheritTWilio(models.Model):
     _inherit = "voip.call"
 
     twilio_sid = fields.Char(string="Twilio SID")
+    twilio_account_id = fields.Many2one('voip.twilio', string="Twilio Account")
+    currency_id = fields.Many2one('res.currency', string="Currency")
+    price = fields.Float(string="Price")
+    margin = fields.Float(string="Margin")
+